[jsroot] pre6  7/12/2020
First support custom credits for WebWindow
Fully change SetDivId - now it is only assign to HTML element
and does not response to create TCanvas or TFrame
Separate menu functionality in extra script
