+++ js/scripts/JSRoot.geobase.js
@@ -2110,7 +2110,7 @@ JSROOT.define(['three', 'csg'], (THREE, ThreeBSP) => {
          source = createProjectionMatrix(source);
 
       let frustum = new THREE.Frustum();
-      frustum.setFromMatrix(source);
+      frustum.setFromProjectionMatrix(source);
 
       frustum.corners = new Float32Array([
           1,  1,  1,
