+++ js/scripts/JSRoot.v7gpad.js
@@ -1424,14 +1424,11 @@ JSROOT.define(['d3', 'painter'], (d3, jsrp) => {
    }
 
    let drawRAxis = (divid, obj /*, opt*/) => {
-
       let painter = new RAxisPainter(obj);
-
-      painter.SetDivId(divid);
-
       painter.disable_zooming = true;
-
-      return painter.Redraw().then(() => painter);
+      return jsrp.ensureRCanvas(painter, divid, false)
+                 .then(() => painter.Redraw())
+                 .then(() => painter);
    }
 
    // ==========================================================================================
@@ -1738,7 +1735,7 @@ JSROOT.define(['d3', 'painter'], (d3, jsrp) => {
    RFramePainter.prototype.DrawAxes = function() {
 
       if (this.axes_drawn || (this.xmin==this.xmax) || (this.ymin==this.ymax))
-         return Promise.resolve(false);
+         return Promise.resolve(this.axes_drawn);
 
       this.CleanupAxes();
 
@@ -2355,7 +2352,7 @@ JSROOT.define(['d3', 'painter'], (d3, jsrp) => {
    RFramePainter.prototype.AddInteractive = function() {
 
       if (JSROOT.BatchMode || (!JSROOT.settings.Zooming && !JSROOT.settings.ContextMenu))
-         return Promise.resolve(false);
+         return Promise.resolve(true);
 
       return JSROOT.require(['interactive']).then(inter => {
          inter.FrameInteractive.assign(this);
@@ -2374,12 +2371,13 @@ JSROOT.define(['d3', 'painter'], (d3, jsrp) => {
       if (handle) handle.ChangeLog('toggle');
    }
 
-   function drawFrame(divid, obj, opt) {
+   function drawRFrame(divid, obj, opt) {
       let p = new RFramePainter(obj);
       if (opt == "3d") p.mode3d = true;
-      p.SetDivId(divid, 2);
-      p.Redraw();
-      return p.DrawingReady();
+      return jsrp.ensureRCanvas(p, divid, false).then(() => {
+         p.Redraw();
+         return p;
+      });
    }
 
    // ===========================================================================
@@ -3519,7 +3517,7 @@ JSROOT.define(['d3', 'painter'], (d3, jsrp) => {
             evnt.stopPropagation();
          }
 
-         if (jsrp.closeMenu()) return;
+         if (jsrp.closeMenu && jsrp.closeMenu()) return;
 
          jsrp.createMenu(this, evnt).then(menu => {
             menu.add("header:Menus");
@@ -4150,13 +4148,10 @@ JSROOT.define(['d3', 'painter'], (d3, jsrp) => {
       return this.has_event_status;
    }
 
-   function drawCanvas(divid, can /*, opt */) {
+   function drawRCanvas(divid, can /*, opt */) {
       let nocanvas = !can;
-      if (nocanvas) {
-         console.log("No canvas specified");
-         return null;
-         // can = JSROOT.Create("ROOT::Experimental::TCanvas");
-      }
+      if (nocanvas)
+         can = JSROOT.Create("ROOT::Experimental::TCanvas");
 
       let painter = new RCanvasPainter(can);
       painter.normal_canvas = !nocanvas;
@@ -4184,6 +4179,34 @@ JSROOT.define(['d3', 'painter'], (d3, jsrp) => {
          return painter;
       });
    }
+
+     /** @summary Ensure TCanvas and TFrame for the painter object
+    * @param {Object} painter  - painter object to process
+    * @param {Object|string} divid - HTML element or element id
+    * @param {string|boolean} frame_kind  - false for no frame or "3d" for special 3D mode
+    * @desc Assign divid, creates TCanvas if necessary, add to list of pad painters and */
+   let ensureRCanvas = function(painter, divid, frame_kind) {
+      if (!painter) return Promise.reject('Painter not provided in ensureRCanvas');
+
+      // assign divid and pad name as required
+      painter.SetDivId(divid, -1);
+
+      // simple check - if canvas there, can use painter
+      let svg_c = painter.svg_canvas();
+      let noframe = (frame_kind === false) || (frame_kind == "3d") ? "noframe" : "";
+
+      let promise = !svg_c.empty() ? Promise.resolve(true) : drawRCanvas(divid, null, noframe);
+
+      return promise.then(() => {
+         if (frame_kind === false) return;
+         if (painter.svg_frame().select(".main_layer").empty())
+            return drawRFrame(divid, null, (typeof frame_kind === "string") ? frame_kind : "");
+      }).then(() => {
+         painter.addToPadPrimitives();
+         return painter;
+      });
+   }
+
 
    // ======================================================================================
 
@@ -4332,15 +4355,13 @@ JSROOT.define(['d3', 'painter'], (d3, jsrp) => {
    let drawPave = (divid, pave, opt) => {
       let painter = new RPavePainter(pave, opt);
 
-      painter.SetDivId(divid);
-
-      return painter.DrawPave();
+      return jsrp.ensureRCanvas(painter, divid, false).then(() => painter.DrawPave());
    }
 
    // =======================================================================================
 
 
-   function drawFrameTitle(reason) {
+   function drawRFrameTitle(reason) {
       let fp = this.frame_painter();
       if (!fp)
          return console.log('no frame painter - no title');
@@ -4756,10 +4777,11 @@ JSROOT.define(['d3', 'painter'], (d3, jsrp) => {
 
    let drawPalette = (divid, palette, opt) => {
       let painter = new RPalettePainter(palette, opt);
-      painter.SetDivId(divid);
-      painter.CreateG(false);
 
-      return painter.DrawingReady();
+      return jsrp.ensureRCanvas(painter, divid, false).then(() => {
+         painter.CreateG(false); // just create container, real drawing will be done by histogram
+         return painter;
+      });
    }
 
    // JSROOT.addDrawFunc({ name: "ROOT::Experimental::RPadDisplayItem", icon: "img_canvas", func: drawPad, opt: "" });
@@ -4768,17 +4790,17 @@ JSROOT.define(['d3', 'painter'], (d3, jsrp) => {
    JSROOT.addDrawFunc({ name: "ROOT::Experimental::RHist2Drawable", icon: "img_histo2d", prereq: "v7hist", func: "JSROOT.v7.drawHist2", opt: "" });
    JSROOT.addDrawFunc({ name: "ROOT::Experimental::RHist3Drawable", icon: "img_histo3d", prereq: "v7hist3d", func: "JSROOT.v7.drawHist3", opt: "" });
    JSROOT.addDrawFunc({ name: "ROOT::Experimental::RHistDisplayItem", icon: "img_histo1d", prereq: "v7hist", func: "JSROOT.v7.drawHistDisplayItem", opt: "" });
-   JSROOT.addDrawFunc({ name: "ROOT::Experimental::RText", icon: "img_text", prereq: "v7more", func: "JSROOT.v7.drawText", opt: "", direct: true, csstype: "text" });
-   JSROOT.addDrawFunc({ name: "ROOT::Experimental::RFrameTitle", icon: "img_text", func: drawFrameTitle, opt: "", direct: true, csstype: "title" });
+   JSROOT.addDrawFunc({ name: "ROOT::Experimental::RText", icon: "img_text", prereq: "v7more", func: "JSROOT.v7.drawText", opt: "", direct: "v7", csstype: "text" });
+   JSROOT.addDrawFunc({ name: "ROOT::Experimental::RFrameTitle", icon: "img_text", func: drawRFrameTitle, opt: "", direct: "v7", csstype: "title" });
    JSROOT.addDrawFunc({ name: "ROOT::Experimental::RPaletteDrawable", icon: "img_text", func: drawPalette, opt: "" });
    JSROOT.addDrawFunc({ name: "ROOT::Experimental::RDisplayHistStat", icon: "img_pavetext", prereq: "v7hist", func: "JSROOT.v7.drawHistStats", opt: "" });
-   JSROOT.addDrawFunc({ name: "ROOT::Experimental::RLine", icon: "img_graph", prereq: "v7more", func: "JSROOT.v7.drawLine", opt: "", direct: true, csstype: "line" });
-   JSROOT.addDrawFunc({ name: "ROOT::Experimental::RBox", icon: "img_graph", prereq: "v7more", func: "JSROOT.v7.drawBox", opt: "", direct: true, csstype: "box" });
-   JSROOT.addDrawFunc({ name: "ROOT::Experimental::RMarker", icon: "img_graph", prereq: "v7more", func: "JSROOT.v7.drawMarker", opt: "", direct: true, csstype: "marker" });
+   JSROOT.addDrawFunc({ name: "ROOT::Experimental::RLine", icon: "img_graph", prereq: "v7more", func: "JSROOT.v7.drawLine", opt: "", direct: "v7", csstype: "line" });
+   JSROOT.addDrawFunc({ name: "ROOT::Experimental::RBox", icon: "img_graph", prereq: "v7more", func: "JSROOT.v7.drawBox", opt: "", direct: "v7", csstype: "box" });
+   JSROOT.addDrawFunc({ name: "ROOT::Experimental::RMarker", icon: "img_graph", prereq: "v7more", func: "JSROOT.v7.drawMarker", opt: "", direct: "v7", csstype: "marker" });
    JSROOT.addDrawFunc({ name: "ROOT::Experimental::RPave", icon: "img_pavetext", func: drawPave, opt: "" });
    JSROOT.addDrawFunc({ name: "ROOT::Experimental::RLegend", icon: "img_graph", prereq: "v7more", func: "JSROOT.v7.drawLegend", opt: "" });
    JSROOT.addDrawFunc({ name: "ROOT::Experimental::RPaveText", icon: "img_pavetext", prereq: "v7more", func: "JSROOT.v7.drawPaveText", opt: "" });
-   JSROOT.addDrawFunc({ name: "ROOT::Experimental::RFrame", icon: "img_frame", func: drawFrame, opt: "" });
+   JSROOT.addDrawFunc({ name: "ROOT::Experimental::RFrame", icon: "img_frame", func: drawRFrame, opt: "" });
    JSROOT.addDrawFunc({ name: "ROOT::Experimental::RAxisDrawable", icon: "img_frame", func: drawRAxis, opt: "" });
    JSROOT.addDrawFunc({ name: "ROOT::Experimental::RAxisLabelsDrawable", icon: "img_frame", func: drawRAxis, opt: "" });
 
@@ -4789,12 +4811,14 @@ JSROOT.define(['d3', 'painter'], (d3, jsrp) => {
    JSROOT.v7.RCanvasPainter = RCanvasPainter;
    JSROOT.v7.RPavePainter = RPavePainter;
    JSROOT.v7.drawRAxis = drawRAxis;
-   JSROOT.v7.drawFrame = drawFrame;
+   JSROOT.v7.drawRFrame = drawRFrame;
    JSROOT.v7.drawPad = drawPad;
-   JSROOT.v7.drawCanvas = drawCanvas;
+   JSROOT.v7.drawRCanvas = drawRCanvas;
    JSROOT.v7.drawPadSnapshot = drawPadSnapshot;
    JSROOT.v7.drawPave = drawPave;
-   JSROOT.v7.drawFrameTitle = drawFrameTitle;
+   JSROOT.v7.drawRFrameTitle = drawRFrameTitle;
+
+   jsrp.ensureRCanvas = ensureRCanvas;
 
    return JSROOT;
